backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
langchain==0.2.2
langchain-core==0.2.38
langchain-community==0.0.35
langchain-pinecone==0.0.3
langgraph==0.3.1
langgraph-checkpoint==2.0.26
pinecone-client==3.0.3
neo4j==5.18.0
python-dotenv==1.0.1
google-generativeai==0.4.0
semantic-kernel==0.9.1b1
